Changed HTML, CSS, and JavaScript for multiple files
login.html -- I changed the styling of the buttons for Student Login and
Teacher Login
password.html -- I added the input to type in the password and submit
button && styled them.
script.js -- I made a function SubmitPassword for when you press the
submit button on password.html. Look at comment in script.js. That is
where enter code to be run when password is correct.