Removed files, Added a file, changed html, styling, and javascript.
I removed 1,2, and 3.html and added student.html for only 1 page for
future purposes. I changed some styling in style.css for my needs on
password.html. I removed login.css for it is not needed anymore. I
edited a lot of javascript so that you can now create and remove codes.
And I set a limit on the amount of codes (1-99).